requirements.txt
fastapi
uvicorn[standard]
motor
jinja2
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
email-validator
black
flake8
pytest
isort
httpx
